requirements.txt
# Prompt Anywhere 依赖
# 安装命令：pip install -r requirements.txt

# 核心框架
PySide6>=6.7

# 全局快捷键
pynput>=1.7.6

# Markdown 渲染
markdown>=3.5
Pygments>=2.16

# 配置读取
PyYAML>=6.0.1

# 环境变量
python-dotenv>=1.0.0
